Three uodates
Updated commands that create and update schemas to convert spaces to `_` in schema and field names as spaces are not valid.

Updated `gam <UserTypeEntity> move drivefile <DriveFileEntity>` to move a folder from a Shared Drive to My Drive by updating the parent rather than creating a new folder.

Upgraded to Python 3.9.1.